Tpetra: Harness: Move withLocalAccess implementation into new header
@trilinos/tpetra

Move withLocalAccess implementation into new header file
Tpetra_withLocalAccess.hpp.  Improve Doxygen documentation.
